Separate surface normalization from dimensionality reduction
The surface normalization code was intermingled with the dimensionality reduction
(i.e. reducing the image height to a region of interest), which caused the need to
create a depth vector and create normalize_surface_vect.m. Now the surface normalization
works on the original image dimensions; this is a little bit slower, but it simplified the
code.

Previously, there were two stages of dimensionality reduction: one during normalize_surface_vect(),
and one after. Now there is only one stage of dimensionality reduction. During the second
dimensionality reduction stage, the surface padding (SURFACE_PAD) was being cut off, so SURFACE_PAD
was changed in to reduce another magic number; now SURFACE_PAD is the only parameter that needs to
be adjusted to adjust where the reduced image starts.

Comments were added to explain the intent behind the data preprocessing steps